Allow a custom SSL socket factory and trust manager to be used.
The client doesn't support an SSL configuration where cert checking is disabled
and turned off. This allows you supply certs via configuration to connect to a
feature server. It's mostly useful for self-signed certificates that aren't on
the system's trust chain.

The path or classpath is scanned for .crt and .pem files. Found certs are
loaded by creating a custom socket factory and trust manager for the
OkHttp client. The path must begin with "file:///" or "classpath:" to indicate
whether the certs are loaded from a file directory or the classpath.