Adds versioning support to the API server.
This allows the server to generate sortable version numbers using a timestamp
and a counter. It's a hacked up impl based on hybrid logical clocks
(https://www.cse.buffalo.edu/tech-reports/2014-04.pdf) with a guard to check
if the local system time has moved "too far" into the future. If it has the
counter is bumped and the last logical time is kept instead.